requirements.txt
alembic==1.10.3
arrow==1.2.3
eventlet==0.30.2
flask-blueprint==1.3.0
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
Flask==2.2.3
gevent-websocket==0.10.1
gevent==22.10.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
numpy==1.24.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-socketio==5.8.0
redis==4.5.4
requests==2.28.2
rq==1.13.0
SQLAlchemy==2.0.10